requirements.txt
Django
django-crispy-forms
django-mptt
